// src/pages/types.ts
export type DashboardItem = {
    name: string;
    link: string;
    id: number;
    icons: string;
    count: number;
}


export type CategoryItem = {
    id: number;
    name: string;
}

export type StatusItem  = {
    id: number;
    name: string;
} 

export type DeveloperItem  = {
    id: number;
    name: string;
} 